// src/js/interfaces/IRootScope.ts
export interface IRootScope extends angular.IRootScopeService {
	messages: Array<any>;
	messageClass: string;
	closeMessage(): void;
	isAuthenticated: boolean;
	sessionId: any;
	fxpUIConstants: any;
	BusinessRole: string;
	roleGroupId: number;
	roleGroupIdInternal: number;
	configRouteStates: Array<any>;
	navigateToLandingPage(): void;
	getLandingPage(): void;
	actOnBehalfOfUserActive: boolean;
	OBOUserRoutes: any;
	currentRoutes: any;
	showLoader: boolean;
	loadingText: string;
	isFxpLoadedWithClaims: boolean;
	pageLoadMetrics: any;
	viewContentLoadingTime: any;
	stateChangeStartTime: any;
	startTime: any;
	isLeftNavPinned: boolean;
	isLeftNavOpen: boolean;
	previousStateName: string;
	fxpBreadcrumb: Array<any>;
	isBreadcrumbEnabled: boolean;
	initialFlags: any;
	displayName: string;
	tenantId: string;
	tenantIdInternal: string;
	isNotificationDNDEnabled: boolean;
	isNewTabAllowed: boolean;
	isHelpOpen: boolean;
	currentLeftNavItem: any;
	isHelpFlyoutPinned: any;
}